preprocess pipeline working

+++ mgs_diff_renderer/src/cuda/mgs_dr_buffers.cu
@@ -11,12 +11,12 @@ MGSDRrenderBuffers::MGSDRrenderBuffers(uint8_t* mem, uint32_t count) :
 MGSDRgeomBuffers::MGSDRgeomBuffers(uint8_t* mem, uint32_t count) :
 	MGSDRrenderBuffers(mem, count)
 {
-	pixCenters = bump<uint2>();
-	pixRadii = bump<float>();
-	depths = bump<float>();
+	pixCenters   = bump<float2>();
+	pixRadii     = bump<float>();
+	depths       = bump<float>();
 	tilesTouched = bump<uint32_t>();
-	conic = bump<float4>();
-	rgb = bump<float3>();
+	conic        = bump<float4>();
+	rgb          = bump<float3>();
 }
 
 MGSDRbinningBuffers::MGSDRbinningBuffers(uint8_t* mem, uint32_t count) :
